feat(natureos): cloud IA May 2026 — app routes, aerosol BFF, tools hub, ancestry move
