+++ app/natureos/nature-statistics/page.tsx
