+++ app/natureos/petri-sim/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation"
 
 export default function NatureOSPetriSimPage() {
-  redirect("/natureos/tools/petri-dish")
+  redirect("/natureos/virtual-petri-dish")
 }
